Start updating pallet to new macros
